move layers request to the component and auto load the first layer
