+++ web/src/components/common/App.tsx
@@ -28,15 +28,12 @@ import { getParsedIframeInfo } from '../../helpers/origin'
 import { getApp } from '../../helpers/app'
 import { getBillingInfo } from '../../app/api/billing'
 import { setBillingInfo } from '../../state/billing/reducer'
-import { setAvailableLayers } from '../../state/semantic-layer/reducer'
 import { SupportButton } from './Support'
 import { Markdown } from './Markdown'
 import { setMinusxMode, toggleMinusXRoot } from '../../app/rpc'
 import { configs } from '../../constants'
-import axios from 'axios'
 import { startNewThread } from '../../state/chat/reducer'
 
-const SEMANTIC_LAYERS_API = `${configs.SEMANTIC_BASE_URL}/layers`
 
 const useAppStore = getApp().useStore()
 
@@ -270,31 +267,6 @@ const AppBody = forwardRef((_props, ref) => {
         console.log('register error is', err)
       })
     }
-  }, [auth.session_jwt])
-  useEffect(() => {
-      const fetchData = async () => {
-          const response = await axios.get(SEMANTIC_LAYERS_API, {
-            headers: {
-              'Content-Type': 'application/json',
-            },
-          })
-          const data = await response.data
-          dispatch(setAvailableLayers(data.layers || []))
-      }
-      if (auth.session_jwt && demoMode) {
-        const MAX_TRIES = 3
-        const tryFetchingSemanticLayer = async (tries = 1) => {
-          if (tries <= MAX_TRIES) {
-            try {
-              await fetchData()
-            } catch (err) {
-              console.warn(`Failed to retrieve semantic properties, try ${tries}`, err)
-              setTimeout(() => tryFetchingSemanticLayer(tries + 1), 1000*tries)
-            }
-          }
-        }
-        tryFetchingSemanticLayer()
-      }
   }, [auth.session_jwt])
   // const IFrameButton = (<IconButton 
   //   borderRightRadius={0} aria-label="Home"
